Refactoring and adding Blazor features

+++ CleanArch.UI/CleanArch.BlazorUI/Pages/LeaveTypes/Index.razor.cs
@@ -1,9 +1,11 @@
+using Blazored.Toast.Services;
 using CleanArch.BlazorUI.Extensions;
 using CleanArch.BlazorUI.Interfaces;
 using CleanArch.BlazorUI.Models.LeaveTypes;
 using CleanArch.BlazorUI.Services.Base;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
+using Microsoft.JSInterop;
 
 namespace CleanArch.BlazorUI.Pages.LeaveTypes;
 
@@ -12,6 +14,8 @@ public partial class Index
     [Inject] private NavigationManager NavigationManager { get; set; } = null!;
     [Inject] private ILeaveTypeService LeaveTypeService { get; set; } = null!;
     [Inject] private ILeaveAllocationService LeaveAllocationService { get; set; } = null!;
+    [Inject] private IToastService ToastService { get; set; } = null!;
+    [Inject] private IJSRuntime Js { get; set; } = null!;
 
     private List<LeaveTypeVM>? LeaveTypes { get; set; }
     public string? Message { get; set; }
@@ -43,15 +47,21 @@ public partial class Index
 
     private async Task DeleteLeaveType(int id)
     {
-        Response<Guid> response = await LeaveTypeService.DeleteLeaveType(id);
+        bool confirm = await Js.InvokeAsync<bool>("confirm", "Do you want to delete this leave type?");
 
-        if(response.Success)
+        if(confirm)
         {
-            StateHasChanged();
-        }
-        else
-        {
-            Message = response.Message;
+            Response<Guid> response = await LeaveTypeService.DeleteLeaveType(id);
+
+            if (response.Success)
+            {
+                ToastService.ShowSuccess("Leave type deleted successfully");
+                await OnInitializedAsync();
+            }
+            else
+            {
+                Message = response.Message;
+            }
         }
     }
 }
